Schema-export now successfully generates the correct tables.
The issue may have been that the "owner" was the User previously,
but now it is the Person who is the owner. That was the biggest change.
However, now I have a problem with a Person not saving after adding
followers. The error is ambiguous as all it states is "error: null".