Added a domain object to join User/Person (followers)
This is the only way I've found so far that actually creates
as I designed it. It might be possible to do what I was doing
before and then use the joinTable in the mappedBy to create
a more customized joinTable. If it is possible, I would prefer
that solution unless there is a major caveat.